movie details to one slice & saga
