+++ src/features/movies/moviesSaga.js
@@ -1,17 +1,21 @@
 import { takeLatest, put, call, delay, select } from "redux-saga/effects";
-import { 
-    fetchError, 
-    fetchMoviesSuccess, 
-    fetchMovies, 
-    fetchGenres, 
-    fetchGenresSuccess, 
-    selectQuery, 
-    setQuery, 
-    selectPage, 
-    setPage, 
-    setLoadingState
+import {
+    fetchError,
+    fetchMoviesSuccess,
+    fetchMovies,
+    fetchGenres,
+    fetchGenresSuccess,
+    fetchMovieDetailsSuccess,
+    fetchMovieCreditsSuccess,
+    fetchMovieDetails,
+    selectQuery,
+    selectId,
+    setQuery,
+    selectPage,
+    setPage,
+    setLoadingState,
 } from "./moviesSlice";
-import { getGenres, getMovies } from "./api";
+import { getGenres, getMovies, getMovieDetails, getMovieCredits } from "./api";
 
 function* fetchMoviesHandler() {
     const query = yield select(selectQuery);
@@ -29,7 +33,7 @@ function* fetchMoviesHandler() {
     } catch (error) {
         yield put(fetchError());
     }
-}
+};
 
 function* fetchGenresHandler() {
     try {
@@ -39,7 +43,20 @@ function* fetchGenresHandler() {
     } catch (error) {
         yield put(fetchError());
     }
-}
+};
+
+function* fetchMovieDetailsHandler() {
+    try {
+        const id = yield select(selectId);
+        const movie = yield call(getMovieDetails, id);
+        const credits = yield call(getMovieCredits, id);
+        yield delay(500);
+        yield put(fetchMovieDetailsSuccess(movie));
+        yield put(fetchMovieCreditsSuccess(credits));
+    } catch (error) {
+        yield put(fetchError());
+    }
+};
 
 function* setOnChangeHandler() {
     yield put(setLoadingState(true));
@@ -47,9 +64,10 @@ function* setOnChangeHandler() {
     yield put(fetchMovies());
 };
 
-export function* watchFetchPopularMovies() {
+export function* moviesSaga() {
+    yield takeLatest(fetchMovieDetails.type, fetchMovieDetailsHandler);
     yield takeLatest(fetchGenres, fetchGenresHandler);
     yield takeLatest(fetchMovies.type, fetchMoviesHandler);
     yield takeLatest(setQuery.type, setOnChangeHandler);
     yield takeLatest(setPage.type, setOnChangeHandler);
-}
+};
