+++ src/features/movies/moviesSlice.js
@@ -5,13 +5,16 @@ const moviesSlice = createSlice({
     initialState: {
         loading: true,
         error: false,
+        query: null,
+        page: 1,
         movies: [],
         genres: [],
-        query: "",
-        page: 1,
+        movieId: "",
+        movieDetails: [],
+        movieCredits: [],
     },
     reducers: {
-        setLoadingState: (state, {payload: boolean}) => {
+        setLoadingState: (state, { payload: boolean }) => {
             state.loading = boolean;
         },
         setQuery: (state, { payload: query }) => {
@@ -37,6 +40,26 @@ const moviesSlice = createSlice({
         fetchGenresSuccess: (state, { payload: genres }) => {
             state.genres = genres;
         },
+        changeMovieId: (state, { payload: id }) => {
+            state.movieId = id;
+        },
+        fetchMovieDetails: (state) => {
+            state.loading = true;
+            state.error = false;
+            state.movieDetails = [];
+        },
+        fetchMovieDetailsSuccess: (state, { payload: movie }) => {
+            state.loading = false;
+            state.movieDetails = movie;
+        },
+        fetchMovieCredits: (state) => {
+            state.loading = true;
+            state.movieCredits = [];
+        },
+        fetchMovieCreditsSuccess: (state, { payload: credits }) => {
+            state.loading = false;
+            state.movieCredits = credits;
+        },
     },
 });
 
@@ -49,7 +72,11 @@ export const {
     fetchMoviesSuccess,
     fetchGenres,
     fetchGenresSuccess,
-
+    changeMovieId,
+    fetchMovieDetails,
+    fetchMovieDetailsSuccess,
+    fetchMovieCredits,
+    fetchMovieCreditsSuccess,
 } = moviesSlice.actions;
 
 export const selectMoviesState = (state) => state.movies;
@@ -59,5 +86,8 @@ export const selectError = (state) => selectMoviesState(state).error;
 export const selectGenres = (state) => selectMoviesState(state).genres;
 export const selectQuery = (state) => selectMoviesState(state).query;
 export const selectPage = (state) => selectMoviesState(state).page;
+export const selectId = (state) => selectMoviesState(state).movieId;
+export const selectMovieDetails = (state) => selectMoviesState(state).movieDetails;
+export const selectMovieCredits = (state) => selectMoviesState(state).movieCredits;
 
 export default moviesSlice.reducer;
